RestTemplateConfig moved to neighbour-service
