+++ neighbour-discoverer/pom.xml
@@ -70,11 +70,6 @@
         </dependency>
 
         <dependency>
-            <groupId>org.apache.httpcomponents</groupId>
-            <artifactId>httpclient</artifactId>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
             <groupId>net.nordicway.interchange</groupId>
             <artifactId>testcontainers-base</artifactId>
             <version>${project.version}</version>
