+++ neighbour-server-app/src/main/java/no/vegvesen/ixn/federation/server/NeighbourRestApiApplication.java
@@ -3,7 +3,7 @@ package no.vegvesen.ixn.federation.server;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication(scanBasePackages = "no.vegvesen.ixn.federation")
+@SpringBootApplication(scanBasePackages = "no.vegvesen.ixn")
 public class NeighbourRestApiApplication {
 
 	public static void main(String[] args) {
